Tutorial Door Finalization
Phased out triggerDoor script in favor of adding three separate scripts that will hopefully work better long term. They check if the player is standing in a particular collider, presses a a keyboard key, adds that 'key' from the terminal to them, and then will check if they have that key or not when they approach a door. For now it will simply make the door collider not active and the sprites disappear. Hoping to beautify and clean it up later.