Saves Implementation
Laid a large amount of the foundations for saving where in the game the player is. Added level 1 and level 2 placeholders to test how it will work.

Level number will be saved to a scriptable object called SaveData from the DataHolder Script. This will be triggered whenever they enter a Level_Transition trigger collider and will both update the level number in the object as well as save that new level number to a JSON file. File location of the JSON is not changeable at the moment without massive changes that aren't really necessary.