Making the UI of the main minigame
Created scene BruteForce_Hashcat for the UI that will be used for the minigame of dragging and dropping the pieces of the hashcat command into the correct spots.

Has two new scripts, DragDrop and CodeSlot. DragDrop is for the item you wish to drag around on the canvas and CodeSlot is for the slot in which you wish the dragged item to snap to when it is overlapping.

For now I have used extremely basic barebones placeholders for these things, but plan to make them look authentic to what we plan to do once I know more about how Hashcat actually works.