Fixed asset overlap, made the Load button functional, and added more tiles and slots to the BruteForce Tutorial Scene
Fixed the asset overlap issue by changing the sorting order of the props vs background vs the player. For some reason it is backwards where default is the very back and background is the very front, so I assigned them accordingly to that.

Load button is functional by using Level_Transition and pulling from the SaveManager script to access the Scriptable object DataHolder to see what level was last accessed by the player and loading them into that specific scene.

Added custom assets to the Brute Scene to make the tiles A,B,C, and D and added more slots accordingly. They all seem to work correctly with the snapping. I will need to add scripts to them to make them actually functional.

TODO: Need to make a bridge between Level_1 and Brute Tutorial, and also need to make a bridge between Tutorial_Level & Level_1 to the main menu. Will likely bind the first set through similar means of what I did with Tutorial and opening the door. For the main menu it will either need a smaller menu to affirm they want to proceed or just take them straight to the main menu at a key press