Made the drag and drop minigame functional
I have both added a new script and modified an existing one to make this work.

The HashcatVerifier has two vital functions that it performs to make this work, it can either Store to the public array it has, or it can Verify against the array of objects that has been deemed the solution for the puzzle. The solution array is determined by a list of objects dragged into it via the Unity GUI, which in this case is the order of which hashcat blocks you wish to be the answer. Currently it's just A,B,C,D but I plan to adjust that later.

The Code Slot script was modified to hold both an index number and a Verifier script for each slot. The script is taken via dragging it over from the Scripts and the Index is to be typed in manually for each slot. This is very important for making sure the Verifier works correctly.