Made UI a prefab, Implemented it into level 1, and made a few tweaks
I made the UI from the BruteForce_Hashcat scene into an actual prefab to be able to port it over to Level 1. I have changed it to where it will start as not active by default and will be made active by the player walking to the first terminal and pressing 'F' at it. I have also made tweaks to this first instance of the prefab by changing the answer to D,C,B,A. For now it will only display a console log saying 'You got it!', but plan to implement functionality where it will enable the user to then interact with the door terminal for the next step. This will hopefully make the implementation of the actual hashcat portion easy as I tried to design the UI to be as modular as I could. In theory it would just require changing the images of the tiles, adding a few extra, and then an implementation of descriptions upon the cursor hovering over each tile.

There is currently a bug where if the user enters from a scene other than level 1 OR closes and attempts to reopen the UI it will take a random amount of multiple attempts of pressing 'F' to do so. I think this may be a weird quirk of the OnTriggerStay function, but will need to do more testing.