Update_Summary
Added a Pause Menu that when pressing "p" will bring up the menu. You can resume the game or save and quit (Save and quit are still in progress)

Made it so that the menu fits the screen in the editor.

Made it so that the player does not move when accessing the menu and when exiting it, they are able to move again.