Made Level_1 fully functional
I added serialized fields to HashcatVerifier where it will grab Gadget_Holder and Terminal_Key from the Player and Terminal objects respectively. They are both used to add various keys to the player once they complete the puzzles in the UI elements. I also made tweaks to the Gadget_Holder script to have it only give a key on 'f' press when they are in the tutorial level. On Level_1 they will only receive the respective keys after they complete the puzzles.

I don't know why the if statement was added *into* the addKey function but I had to move it because it was forcing me to get null pointer exceptions for trying to make this work because the objective board stuff is not anywhere other than the tutorial. Please do not add hyper specific things like that to my functions without telling me about it in the future.

My future plans include fixing the issue where you need to spam 'f' to open some UIs. I have some ideas on how I want to fix this and will either do so tonight or sometime tomorrow. I also want to change Level_2 to either a completion screen or something other than an empty sad void. I also plan on adding more variables to the hashcat command and the hover description function.