Fixed the UI issue with it not always opening on key press
So essentially the issue is that OnTriggerStay2D does *not* check every moment that the game is running like Update() does. To fix this I implemented a slightly hacky solution where I put two if statements into an Update() function. One of them will check if the 'F' key is being pressed. If it is then a specific boolean will flip to true. If it is no longer being pressed, it will flip back to false. I am now only checking for this boolean in OnTriggerStay2D instead of checking for the key press itself. This seems to have entirely resolved the issue of the menus not opening instantly or consistently. Further testing will be required to make sure this is working correctly.