Created functionality for the description to change with cursor position
I created a new script that creates an array of structs where you can tell it what tile ties to what description and then it will check to see which description it should be displaying at a given time. I have added descriptions for all the hashcat tiles and added the a3 tile as well. Made sure to comment everything accordingly to explain how it all works. Overall the full functionality is almost done.

Main things that still need tweaked are implementing a visual change when they complete the puzzle correctly and then doing a bit of cleanup on the final level to not look so bare bones.