Added Victory Messages and Major Refactoring
When the player completes the tutorial and the hashcat menu it will disable the generic description and replace it with a congratulations message when they get it correct.

I also fixed the Load button to ensure it works correctly every time and will reset the load to level 1 after they are taken back to the main menu upon completing the game.

I removed triggerDoor and LevelManager as they were redundant and no longer needed for any game functionalities. I have changed the script used for the Pause menu to Level_Transition and tested to ensure it works for all levels.

The game should now me fully functional from start to end and be able to be replayed however many times as needed. I will be doing one more push after this to do a more widespread purge to remove all unneeded files and folders that are still in the Assets.